Move TutorialPageTwo into its own file
